fix(agent): 修复 session_search clamp 类型与 remember 更新保护，补齐单测

+++ lib/services/agent_tools/session_search_tool.dart
@@ -75,8 +75,10 @@ class SessionSearchTool extends AgentTool {
       );
     }
 
-    final limit =
-        call.getInt('limit', defaultValue: defaultLimit).clamp(1, maxLimit);
+    final limit = call
+        .getInt('limit', defaultValue: defaultLimit)
+        .clamp(1, maxLimit)
+        .toInt();
 
     try {
       final now = DateTime.now();
